Transfer test case from defer to go statement
